Move generics tests to extra file & test pymethods
